+++ src/private/release/GetNovaVersionUpdatePlan.ps1
@@ -11,7 +11,7 @@ function Get-NovaVersionUpdatePlan {
     $jsonContent = Get-Content -LiteralPath $projectInfo.ProjectJSON -Raw | ConvertFrom-Json
     [semver]$currentVersion = $jsonContent.Version
     $versionPart = Get-NovaVersionPartForLabel -CurrentVersion $currentVersion -Label $Label
-    $releaseType = Get-NovaVersionPreReleaseLabel -CurrentVersion $currentVersion -PreviewRelease:$PreviewRelease -StableRelease:$StableRelease
+    $releaseType = Get-NovaVersionPreReleaseLabel -PreviewRelease:$PreviewRelease -StableRelease:$StableRelease
     $newVersion = [semver]::new($versionPart.Major, $versionPart.Minor, $versionPart.Patch, $releaseType, $null)
 
     return [pscustomobject]@{